used double instead of float

+++ WeatherTest/WeatherTest/dataParsing.swift
@@ -40,8 +40,8 @@ struct Response: Codable{
 }
 
 struct Current: Codable{
-    var temp_f : Float = 0
-    var temp_c : Float = 0
+    var temp_f : Double = 0
+    var temp_c : Double = 0
     var condition : Condition?
 }
 struct Condition: Codable{
@@ -58,18 +58,18 @@ struct Forecastday: Codable{
 }
 
 struct Day: Codable{
-    var maxtemp_f : Float = 0
-    var mintemp_f : Float = 0
-    var avgtemp_f : Float = 0
+    var maxtemp_f : Double = 0
+    var mintemp_f : Double = 0
+    var avgtemp_f : Double = 0
     var condition : Condition?
 }
 
 struct Hour: Codable{
     var time : String = "2022-01-04 00:00"
-    var temp_f : Float = 0
+    var temp_f : Double = 0
     var condition : Condition?
-    var wind_mph : Float = 0
-    var feelslike_f : Float = 0
+    var wind_mph : Double = 0
+    var feelslike_f : Double = 0
 }
 extension Hour: Identifiable{
     var id: String {return time}
